+++ datanode/sqlstore/teams.go
@@ -22,14 +22,15 @@ import (
 
 	"code.vegaprotocol.io/vega/datanode/entities"
 	"code.vegaprotocol.io/vega/datanode/metrics"
+	"code.vegaprotocol.io/vega/libs/num"
 	v2 "code.vegaprotocol.io/vega/protos/data-node/api/v2"
+	eventspb "code.vegaprotocol.io/vega/protos/vega/events/v1"
 
 	"github.com/georgysavva/scany/pgxscan"
 )
 
 const (
-	listTeamsStatsQuery = `
-WITH
+	listTeamsStatsQuery = `WITH
   -- This CTE retrieves the all teams statistics reported for the last N epochs.
   eligible_stats AS (
     SELECT *
@@ -39,10 +40,12 @@ WITH
       FROM teams_stats
     ) %s
   ),
-  team_rewards AS (
+  team_numbers AS (
     SELECT t.team_id,
-           SUM(total_quantum_reward) AS total_in_quantum,
-           JSONB_AGG(JSONB_BUILD_OBJECT('epoch', at_epoch, 'total', total_quantum_reward) ORDER BY at_epoch, total_quantum_reward) AS list
+           SUM(total_quantum_reward) AS total_quantum_rewards,
+           JSONB_AGG(JSONB_BUILD_OBJECT('epoch', at_epoch, 'total', total_quantum_reward) ORDER BY at_epoch, total_quantum_reward) AS quantum_rewards,
+           SUM(total_quantum_volume) AS total_quantum_volumes,
+           JSONB_AGG(JSONB_BUILD_OBJECT('epoch', at_epoch, 'total', total_quantum_volume) ORDER BY at_epoch, total_quantum_volume) AS quantum_volumes
     FROM eligible_stats t
     GROUP BY
       t.team_id
@@ -57,17 +60,18 @@ WITH
     GROUP BY
       team_id
   )
-SELECT mr.team_id AS team_id,
-       mr.total_in_quantum AS total_quantum_rewards,
-       mr.list AS quantum_rewards,
+SELECT tn.team_id AS team_id,
+       tn.total_quantum_rewards AS total_quantum_rewards,
+       tn.quantum_rewards AS quantum_rewards,
+       tn.total_quantum_volumes AS total_quantum_volumes,
+       tn.quantum_volumes AS quantum_volumes,
        mg.list AS games_played,
        mg.count AS total_games_played
-FROM team_rewards mr
-  LEFT JOIN team_games mg ON mr.team_id = mg.team_id
+FROM team_numbers tn
+  LEFT JOIN team_games mg ON tn.team_id = mg.team_id
 `
 
-	listTeamMembersStatsQuery = `
-WITH
+	listTeamMembersStatsQuery = `WITH
   -- This CTE retrieves the all teams statistics reported for the last N epochs.
   eligible_stats AS (
     SELECT *
@@ -77,11 +81,13 @@ WITH
       FROM teams_stats
     ) AND team_id = $2 %s
   ),
-  members_rewards AS (
+  members_numbers AS (
     SELECT team_id,
            party_id,
-           SUM(total_quantum_reward) AS total_in_quantum,
-           JSONB_AGG(JSONB_BUILD_OBJECT('epoch', at_epoch, 'total', total_quantum_reward) ORDER BY at_epoch, total_quantum_reward) AS quantum_rewards
+           SUM(total_quantum_reward) AS total_quantum_rewards,
+           JSONB_AGG(JSONB_BUILD_OBJECT('epoch', at_epoch, 'total', total_quantum_reward) ORDER BY at_epoch, total_quantum_reward) AS quantum_rewards,
+           SUM(total_quantum_volume) AS total_quantum_volumes,
+           JSONB_AGG(JSONB_BUILD_OBJECT('epoch', at_epoch, 'total', total_quantum_volume) ORDER BY at_epoch, total_quantum_volume) AS quantum_volumes
     FROM eligible_stats
     GROUP BY
       team_id,
@@ -99,13 +105,22 @@ WITH
       team_id,
       party_id
   )
-SELECT mr.party_id AS party_id,
-       mr.total_in_quantum AS total_quantum_rewards,
-       mr.quantum_rewards AS quantum_rewards,
+SELECT mn.party_id AS party_id,
+       mn.total_quantum_rewards AS total_quantum_rewards,
+       mn.quantum_rewards AS quantum_rewards,
+       mn.total_quantum_volumes AS total_quantum_volumes,
+       mn.quantum_volumes AS quantum_volumes,
        mg.list AS games_played,
        mg.count AS total_games_played
-FROM members_rewards mr
-  LEFT JOIN members_games mg ON mr.team_id = mg.team_id AND mr.party_id = mg.party_id`
+FROM members_numbers mn
+  LEFT JOIN members_games mg ON mn.team_id = mg.team_id AND mn.party_id = mg.party_id`
+
+	upsertTeamsStats = `INSERT INTO teams_stats(team_id, party_id, at_epoch, total_quantum_volume, total_quantum_reward, games_played)
+VALUES
+  %s
+ON CONFLICT (team_id, party_id, at_epoch) DO UPDATE
+  SET total_quantum_volume = excluded.total_quantum_volume
+	`
 )
 
 type (
@@ -244,6 +259,41 @@ func (t *Teams) RefereeSwitchedTeam(ctx context.Context, referee *entities.Refer
 	)
 
 	return err
+}
+
+func (t *Teams) TeamsStatsUpdated(ctx context.Context, evt *eventspb.TeamsStatsUpdated) error {
+	defer metrics.StartSQLQuery("Teams", "TeamsStatsUpdated")()
+
+	var args []interface{}
+
+	values := []string{}
+	for _, teamStats := range evt.Stats {
+		for _, memberStats := range teamStats.MembersStats {
+			notionalVolume, hasErr := num.UintFromString(memberStats.NotionalVolume, 10)
+			if hasErr {
+				notionalVolume = num.UintZero()
+			}
+
+			values = append(values, fmt.Sprintf("(%s, %s, %s, %s, 0, '{}'::JSONB)",
+				nextBindVar(&args, entities.TeamID(teamStats.TeamId)),
+				nextBindVar(&args, entities.PartyID(memberStats.PartyId)),
+				nextBindVar(&args, evt.AtEpoch),
+				nextBindVar(&args, notionalVolume)),
+			)
+		}
+	}
+
+	if len(values) == 0 {
+		return nil
+	}
+
+	query := fmt.Sprintf(upsertTeamsStats, strings.Join(values, ","))
+	_, err := t.Connection.Exec(ctx, query, args...)
+	if err != nil {
+		return fmt.Errorf("could not insert team stats update: %w", err)
+	}
+
+	return nil
 }
 
 func (t *Teams) GetTeam(ctx context.Context, teamID entities.TeamID, partyID entities.PartyID) (*entities.Team, error) {